Move compat into l.lib.
