+++ landscape/package/facade.py
@@ -18,7 +18,7 @@ from apt.progress.base import InstallProgress
 from twisted.python.compat import itervalues
 
 
-from landscape.compat import StringIO
+from landscape.lib.compat import StringIO
 from landscape.lib.fs import append_text_file, create_text_file
 from landscape.lib.fs import read_text_file, read_binary_file, touch_file
 from landscape.package.skeleton import build_skeleton_apt
