+++ landscape/sysinfo/tests/test_sysinfo.py
@@ -4,7 +4,7 @@ import os
 
 from twisted.internet.defer import Deferred, succeed, fail
 
-from landscape.compat import StringIO
+from landscape.lib.compat import StringIO
 from landscape.plugin import PluginRegistry
 from landscape.sysinfo.sysinfo import SysInfoPluginRegistry, format_sysinfo
 from landscape.tests.helpers import LandscapeTest
